// front-end/src/store/user/action-types.ts
export enum UserActionTypes {
  ACTION_LOGIN_BY_WECHAT_CODE = 'ACTION_LOGIN_BY_WECHAT_CODE',
  ACTION_LOGIN_BY_MOBILE_CODE = 'ACTION_LOGIN_BY_MOBILE_CODE',
  ACTION_LOGIN_BY_MOBILE_PASS = 'ACTION_LOGIN_BY_MOBILE_PASS',
  ACTION_RESET_TOKEN = 'ACTION_RESET_TOKEN',
  ACTION_GET_USER_INFO = 'ACTION_GET_USER_INFO',
  ACTION_LOGIN_OUT = 'ACTION_LOGIN_OUT',
  ACTION_REGISTER = 'ACTION_REGISTER',
}